lto: Use C version for SYSCALL_ALIAS / cond_syscall
With LTO aliases get largely resolved in the compiler, not in the linker.

Implement cond_syscall and SYSCALL_ALIAS in C to let the compiler
understand the aliases so that it can resolve it properly.

Likely the architecture specific versions are now not needed
anymore, but I kept them for now.

There is one sublety here:

The assembler version didn't care whether there was a prototype
or not. This variant assumes there is no prototypes because
it uses a dummy (void) signature. This works for sys_ni.c,
but breaks for kernel/time/posix-stubs.c. To avoid problems
here I added a second variant of the macro (_PROTO) that
uses the previous type.

I actually tried to avoid this by adding prototypes for
sys_ni and only use the _PROTO variant, but it resulted in very large
patches and lots of problems with all the different cases.
Eventually I gave up and just special cases the prototype
case in posix-stubs.c

[js] gcc >= 8 emits Wattribute-alias warning. Fix that.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Martin Liska <mliska@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>