x86/lto: Mark 32bit mem{cpy,move,set} __used
gcc doesn't always recognize that memcpy/set/move called
through __builtins are referenced because the reference happens too
late in the RTL expansion phase. This can lead LTO to drop
them, leading to undefined symbols. Mark them as __used to avoid
that.

This is only needed on 32bit, on 64bit they're assembler anyways.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Martin Liska <mliska@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>